tests/generators: Move interfaces and calculation functions to separate module
